Max. time interval error

Maximum time interval error

Formula

Maximum time interval error

\[Mtie(\tau)=\operatorname{max}_{1\leq k\leq N-n}\left(\operatorname{max}_{k\leq t\leq k+n}(x_t)-\operatorname{min}_{k\leq t\leq k+n}(x_t)\right)\]

Doc String

AllanDeviations.mtie — Method.

mtie(data, rate; [frequency=false], [overlapping=true], [taus=Octave]) Calculates the maximal time interval error

parameters:

<data>: The data array to calculate the deviation from either as as phases or frequencies.

<rate>: The rate of the data given.

[frequency]: True if data contains frequency data otherwise (default) phase data is assumed.

[overlapping]: True (default) to calculate overlapping deviation, false otherwise.

[taus]: Taus to calculate the deviation at. This can either be an AllanTauDescriptor type AllTaus, Decadade, HalfDecade, Octave, HalfOctave, QuarterOctave, an array of taus to calculate at, a float number to build a custom log-scale on or an integer to build a specific number of log spaced points.

returns: named tupple (tau, deviation, error, count)

tau: Taus which where used.

deviation: Deviations calculated.

error: Respective errors.

count: Number of contributing terms for each deviation.

source

Possible issues

mtie in itself needs a great amount of computations and can be very slow for big taus with many data points. When computations need too much time, consider reducing the number of taus and/or especially using smaller taus.

Mtie can be called with a non-overlapping calculation. This throws a warning because it is unusual to use but nevertheless faster.