Total Deviation

Total deviation

Formula

total variance

\[Tot\,\sigma_y^2(\tau)=\frac{1}{2\tau^2(N-2)}\sum_{j=2}^{N-1}\left(x_{j-m}^*-2x_{j}^*+x_{j+m}^*\right)^2\]

Doc String

AllanDeviations.totaldev — Method.

totaldev(data, rate; [frequency=false], [overlapping=true], [taus=Octave]) Calculates the total deviation

#parameters:

<data>: The data array to calculate the deviation from either as as phases or frequencies.

<rate>: The rate of the data given.

[frequency]: True if data contains frequency data otherwise (default) phase data is assumed.

[overlapping]: True (default) to calculate overlapping deviation, false otherwise.

[taus]: Taus to calculate the deviation at. This can either be an AllanTauDescriptor type AllTaus, Decadade, HalfDecade, Octave, HalfOctave, QuarterOctave, an array of taus to calculate at, a float number to build a custom log-scale on or an integer to build a specific number of log spaced points.

#returns: named tupple (tau, deviation, error, count)

tau: Taus which where used.

deviation: Deviations calculated.

error: Respective errors.

count: Number of contributing terms for each deviation.

source

Possible issues

totaldev can be called with a non-overlapping calculation. This throws a warning because it is unusual to use but nevertheless faster.